+++ CHANGELOG.md
@@ -25,7 +25,40 @@ The functionalities of each part, Broker and Extensions, on this release, is the
 
 #### Patch Changes
 
+- Overhaul of the Postman-Collection
+
 ### Deployment Migration Notes
+
+## [7.5.0] - 2024-05-16
+
+### Overview
+
+Additional Wrapper API features
+
+### EDC UI
+
+- https://github.com/sovity/edc-ui/releases/tag/v3.2.2
+
+### EDC Extensions
+
+#### Minor Changes
+
+- API Wrapper Use Case API: Catalog endpoint
+
+#### Patch Changes
+
+Security updates
+
+### Deployment Migration Notes
+
+#### Compatible Versions
+
+- Connector Backend Docker Images:
+    - Dev EDC: `ghcr.io/sovity/edc-dev:7.5.0`
+    - sovity EDC CE: `ghcr.io/sovity/edc-ce:7.5.0`
+    - MDS EDC CE: `ghcr.io/sovity/edc-ce-mds:7.5.0`
+- Connector UI Docker Image: `ghcr.io/sovity/edc-ui:3.2.2`
+- Connector UI Release: https://github.com/sovity/edc-ui/releases/tag/v3.2.2
 
 ## [7.4.2] - 2024-04-20